pyproject.toml
[tool.poetry]
name = "logtrucks"
version = "0.1.0"
description = ""
authors = ["Bogdan Despotov <despotov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.2.0"
pyyaml = "^6.0"
pillow = "^9.4.0"
opencv-python = "3.4.11.43"
tensorflow = "2.2"
protobuf = "3.20"
numpy = "1.20.3"
sqlalchemy = "1.4.41"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
